+++ src/pybamm/__init__.py
@@ -107,8 +107,7 @@ from .models.submodels.interface import open_circuit_potential
 # Geometry and domains
 from .geometry.geometry import Geometry
 from .geometry.battery_geometry import battery_geometry
-from .geometry.domains import Domain, Domain1D
-from .expression_tree.independent_variable import KNOWN_COORD_SYS
+from .geometry.domain import Domain
 from .geometry import standard_spatial_vars
 
 # Parameter classes and methods