+++ tests/unit/test_expression_tree/test_independent_variable.py
@@ -35,30 +35,13 @@ class TestIndependentVariable:
         assert t.evaluate_for_shape() == 0
 
     def test_spatial_variable(self):
-        x = pybamm.SpatialVariable("x", "negative electrode")
-        assert x.name == "x"
+        x = pybamm.SpatialVariable("negative electrode")
         assert not x.evaluates_on_edges("primary")
-        y = pybamm.SpatialVariable("y", "separator")
-        assert y.name == "y"
-        z = pybamm.SpatialVariable("z", "positive electrode")
-        assert z.name == "z"
-        r = pybamm.SpatialVariable("r", "negative particle")
-        assert r.name == "r"
         with pytest.raises(NotImplementedError):
             x.evaluate()
 
-        with pytest.raises(ValueError, match="domain must be"):
-            pybamm.SpatialVariable("x", [])
-        with pytest.raises(pybamm.DomainError):
-            pybamm.SpatialVariable("r_n", ["positive particle"])
-        with pytest.raises(pybamm.DomainError):
-            pybamm.SpatialVariable("r_p", ["negative particle"])
-        with pytest.raises(pybamm.DomainError):
-            pybamm.SpatialVariable("x", ["negative particle"])
-
     def test_spatial_variable_edge(self):
-        x = pybamm.SpatialVariableEdge("x", "negative electrode")
-        assert x.name == "x"
+        x = pybamm.SpatialVariableEdge("negative electrode")
         assert x.evaluates_on_edges("primary")
 
     def test_to_equation(self):